extracted upd_server into separate component
