+++ components/udp_server/CMakeLists.txt
@@ -1,0 +1,4 @@
+idf_component_register(SRCS "udp_server.c"
+                       INCLUDE_DIRS "."
+                       REQUIRES lwip
+                       PRIV_REQUIRES esp_event log)
